Move error type to its own module
